backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
fsrs==4.1.1
elevenlabs==1.3.0
google-generativeai==0.3.2

